requirements.txt
asgiref==3.7.2
astroid==2.15.5
black==23.3.0
click==8.1.3
dill==0.3.6
Django==4.2.2
django-restframework==0.0.1
djangorestframework==3.14.0
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pylint==2.17.4
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.11.8
wrapt==1.15.0
